+++ tests/rust/integration/golden/corpus/ontime_flights/test_pattern_matrix__TestDenormalizedSchema_test_graph_funcs_0.databricks.sql
@@ -1,5 +1,5 @@
 SELECT 
-      'FLIGHT::Airport::Airport' AS `type(r)`, 
+      'FLIGHT' AS `type(r)`, 
       r.Origin AS `id(a)`, 
       array('Airport') AS `labels(a)`
 FROM default.flights AS r